New objects, upgrade to Java 8, and more
Lots of progress over the last couple of hours. I was looking into using
a local MySQL database (it would have everything - rooms linked with
pictures and other sweet functionality), but it seems like more trouble
than its worth for such a small program and nearing deadlines. The main
objects I've created are Rooms, Reservations, and the Database (should
probably make that static). I plan to add Users and other goodies as
well in the future. I did move to 1.8, which allows me to use updated
Dates and Periods, which will help with my internal registration system.
There is so much I would do if I had more time, but will have to make
compromises to expedite this thing (I had been coding on a school
project for like 8 hours straight before starting this today)... At the
beginning of the project, I envisioned a Canvas of sorts that would
allow some interactive maps, pics, and selection of rooms. That's not
happening... Which reminds me, I will start recording man-hours put into
this project so I can get some idea of efficiency. I started at roughly
5:30PM EST, so I've put around 3-4 hours into this commit. Yesterday, I
put another two-ish, but it mostly comprised of planning and fooling
around with the GUI. Also, note the addition of the runnable jar. I will
have to remind myself to update it with each commit. Ant is giving me
fits when I try to build one. I've tried scripting it on three different
occasions now...