Added improved date selection
The new selection features both starting and ending date, initializing
with previously entered values. I've cleaned up the old code with some
getters and setters. Input sanitation for the dates is minimal. Outside
being valid and functional objects, there are no guarantees that they
are reasonable (start before end, both in the future, not 1000 years
ahead, etc.). It should be noted that the month ComboBox in the
dateSelector starts at 0. I recognize this and am not going to delve
into that code to fix it. JAR has been updated.  I spent around an hour
on this commit.