Added list updates and query helpers
The lists are now interactive, and the foundation is set for me to
create some more complex SQL statements (only select rooms that lack
reservations that conflict with the following dates). After that, I will
add some descriptions of the selected rooms, have changing images, etc.
The one thing that really held me up in this commit was troubleshooting
why the drivers weren't seen within the JAR. Finally, I simply moved
them out of my library and imported them as separate jars. Mainly, I'm
just surprised that it was an issue. I thought I had some SQL statements
and DB functionality in the last build, so it was really unexpected for
it to become an issue in this commit. I've been on-and-off in the last
hour or so.