Updated updateLists function
The function now updates description, price, and name. Also, I set a
minimum GUI size and did some other light work to it. As usual, the JAR
has been updated to reflect the changes. Next on my to-do list is to add
picture functionality. After that, its User login back-end and finally
check-out and creating new reservations in the DB.