Added all pictures, minor changes to updatePics()
The pictures are now added, and I don't intend to add any more any time
soon. This should be enough to give the viewer an idea of the
functionality. As far as the actual loading, it could be optimized
pretty well, but right now that is more of a minor nuisance than an
actual goal/objective. I created a package for the pics, and changed the
tab numbers so they go from 1-infinity, not 0-infinity. Jar updated.