Finished Reservation System, polished GUI listeners, etc.
The program should be idiot-proof on the outside, despite rather fragile
construction on the internals. I don't plan to fix a lot of those
issues, but rather improve in future projects. Spent about 2 hours
on-and-off working on this. Next, I'm going to tooltip GUI components,
but the program is officially 100% functional for its intended purpose.